Clean up EncodingUtils
